tutorial-production-ai-reasoning/app/main.py
"""
Main FastAPI application for production AI reasoning systems.

This application provides intelligent routing of AI reasoning requests
with cost optimization and monitoring.
"""

from fastapi import FastAPI, HTTPException, BackgroundTasks
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import Optional
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Initialize FastAPI app
app = FastAPI(
    title="Production AI Reasoning System",
    description="Enterprise-grade AI reasoning with cost optimization",
    version="1.0.0"
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Request/Response models
class ReasoningRequest(BaseModel):
    prompt: str
    strategy: Optional[str] = "auto"  # auto, direct, chain, tree
    force_reasoning: bool = False

class ReasoningResponse(BaseModel):
    strategy_used: str
    final_answer: str
    total_tokens: int
    processing_time_ms: int
    cost_usd: float

# API endpoints
@app.post("/api/v1/reason", response_model=ReasoningResponse)
async def reason(request: ReasoningRequest):
    """
    Main reasoning endpoint with automatic strategy selection.
    """
    try:
        # Placeholder implementation
        # In production, this would use the classifier and reasoning engine
        return ReasoningResponse(
            strategy_used=request.strategy if request.strategy != "auto" else "direct",
            final_answer="This is a placeholder response. Implement the full reasoning logic from the tutorial article.",
            total_tokens=150,
            processing_time_ms=800,
            cost_usd=0.002
        )
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/api/v1/metrics")
async def get_metrics(hours: int = 24):
    """Get cost and performance metrics."""
    return {
        "total_requests": 0,
        "total_tokens": 0,
        "total_cost_usd": 0.0,
        "average_cost_per_request": 0.0
    }

@app.get("/api/v1/recommendations")
async def get_recommendations():
    """Get optimization recommendations."""
    return {
        "recommendations": ["System running optimally"]
    }

@app.get("/health")
async def health_check():
    """Health check endpoint."""
    return {
        "status": "healthy",
        "version": "1.0.0"
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
